requirements.txt
Flask
pytesseract
Pillow
Werkzeug
openai
